feat(scribo): update

See merge request olena/pylene!147

+++ pylene/src/scribo/segdet/process/linearregression.hpp
@@ -27,7 +27,7 @@ namespace scribo::internal
   public:
     MemoryLinearRegressor(float x0, float y0, int memory = 20);
     float predict();
-    void  push(size_t x, size_t y);
+    void  push(float x, float y);
   };
 
 } // namespace scribo::internal
